:bug: Fixed two bugs in the route tree creation
The first bug was that the absolut filepaths of the routes where not absolut but relative which created problems in the route registration.
The second bug was that directories that start with '_mw' where added to the routeTree as empty routes wich also caused problems in the route registartion. This route will be registerd as //example